refactor(api): C4 — PollRuntime (api/poll.py) + health module (api/health.py) (ADR-0015)
main.py sheds its two biggest non-endpoint regions (~380 lines):

- NEW api/poll.py: PollRuntime owns the poll machinery — durable ownership
  claims (OwnerStore), per-instance segments, poll_degraded, last_poll_at,
  the shared lock + cycle counter, and the missions cache (ONE stable list
  identity, served by reference). Constructed once in the composition root
  with LIVE references (managers dict, config, forge runtime); never rebuilt
  on config reload — hot-swapped instances appear because build_managers()
  mutates the shared dict in place (ADR-0015 Decision 2). _claim_missions
  stays a pure module function. Bodies moved verbatim (globals → self).
- NEW api/health.py: the four probes (+ their caches), reset_protection_cache
  (config-reload/secret-write hook), and build_health_payload(...) with every
  dependency explicit. /health in main.py is now a one-line forward; the two
  secret routes' inline cache pokes go through reset_protection_cache().
- main.py: poll_rt construction replaces the global block; lifespan spawns
  poll_rt.loop(); /missions, /poll/run, mission-action forwards, and
  clear-runs rewired onto poll_rt. force_poll_now's existing injection seam
  (lock/next_cycle_id/run_cycle) fits PollRuntime unchanged.
- tests: test_multi_instance's four poll tests now construct a PollRuntime
  over fakes (_rt helper) instead of monkeypatching app_main module globals —
  net-better isolation; poll_instance overridden as an instance attribute
  (same pattern as fakes' _audit). _claim_missions imports move to api.poll.

Verified: ruff clean; full suite 637 passed / 1 skipped via pytest_app.sh.
Live /health shape check rides the post-C6 deploy checkpoint.
